enhance: support inline answer collection creation for selection and case study questions in templates (#4709)
